+++ Modules/Intent.Modules.Common/Templates/StaticContent/StaticContentTemplate.cs
@@ -1,3 +1,4 @@
+#nullable enable
 using System;
 using System.Collections.Generic;
 using System.IO;
@@ -14,13 +15,14 @@ namespace Intent.Modules.Common.Templates.StaticContent
     public class StaticContentTemplate : IntentTemplateBase
     {
         private readonly string _sourcePath;
-        private readonly string _relativeOutputPathPrefix;
+        private readonly string? _relativeOutputPathPrefix;
         private readonly OverwriteBehaviour _overwriteBehaviour;
+        private readonly Func<ITemplateFileConfig, StaticContentTemplate, ITemplateFileConfig> _fileConfigConfigurationUpdater;
         private readonly IReadOnlyDictionary<string, string> _replacements;
         private readonly string _relativeOutputPath;
 
         /// <summary>
-        /// Obsolete. Use <see cref="StaticContentTemplate(string,string,string,string,IOutputTarget,IReadOnlyDictionary{string,string},OverwriteBehaviour)"/> instead.
+        /// Obsolete. Use <see cref="StaticContentTemplate(string,string,string,string,IOutputTarget,IReadOnlyDictionary{string,string},OverwriteBehaviour,Func{ITemplateFileConfig,ITemplateFileConfig})"/> instead.
         /// </summary>
         [Obsolete(WillBeRemovedIn.Version4)]
         public StaticContentTemplate(
@@ -36,12 +38,13 @@ namespace Intent.Modules.Common.Templates.StaticContent
                 templateId: templateId,
                 outputTarget: outputTarget,
                 replacements: replacements,
-                overwriteBehaviour: OverwriteBehaviour.OverwriteDisabled)
+                overwriteBehaviour: OverwriteBehaviour.OverwriteDisabled,
+                fileConfigConfigurationUpdater: (fileConfig, _) => fileConfig)
         {
         }
 
         /// <summary>
-        /// Obsolete. Use <see cref="StaticContentTemplate(string,string,string,string,IOutputTarget,IReadOnlyDictionary{string,string},OverwriteBehaviour)"/> instead.
+        /// Obsolete. Use <see cref="StaticContentTemplate(string,string,string,string,IOutputTarget,IReadOnlyDictionary{string,string},OverwriteBehaviour,Func{ITemplateFileConfig,ITemplateFileConfig})"/> instead.
         /// </summary>
         [Obsolete(WillBeRemovedIn.Version4)]
         public StaticContentTemplate(
@@ -58,12 +61,15 @@ namespace Intent.Modules.Common.Templates.StaticContent
                 templateId: templateId,
                 outputTarget: outputTarget,
                 replacements: replacements,
-                overwriteBehaviour: overwriteBehaviour)
+                overwriteBehaviour: overwriteBehaviour,
+                fileConfigConfigurationUpdater: (fileConfig, _) => fileConfig)
         {
         }
+
         /// <summary>
-        /// Creates a new instance of <see cref="StaticContentTemplate"/>.
+        /// Obsolete. Use <see cref="StaticContentTemplate(string,string,string,string,IOutputTarget,IReadOnlyDictionary{string,string},OverwriteBehaviour,Func{ITemplateFileConfig,ITemplateFileConfig})"/> instead.
         /// </summary>
+        [Obsolete(WillBeRemovedIn.Version4)]
         public StaticContentTemplate(
             string sourcePath,
             string relativeOutputPath,
@@ -71,10 +77,35 @@ namespace Intent.Modules.Common.Templates.StaticContent
             string templateId,
             IOutputTarget outputTarget,
             IReadOnlyDictionary<string, string> replacements,
-            OverwriteBehaviour overwriteBehaviour) : base(templateId, outputTarget)
+            OverwriteBehaviour overwriteBehaviour)
+            : this(
+                sourcePath: sourcePath,
+                relativeOutputPath: relativeOutputPath,
+                relativeOutputPathPrefix: relativeOutputPathPrefix,
+                templateId: templateId,
+                outputTarget: outputTarget,
+                replacements: replacements,
+                overwriteBehaviour: overwriteBehaviour,
+                fileConfigConfigurationUpdater: (fileConfig, _) => fileConfig)
+        {
+        }
+
+        /// <summary>
+        /// Creates a new instance of <see cref="StaticContentTemplate"/>.
+        /// </summary>
+        public StaticContentTemplate(
+            string sourcePath,
+            string relativeOutputPath,
+            string? relativeOutputPathPrefix,
+            string templateId,
+            IOutputTarget outputTarget,
+            IReadOnlyDictionary<string, string> replacements,
+            OverwriteBehaviour overwriteBehaviour,
+            Func<ITemplateFileConfig, StaticContentTemplate, ITemplateFileConfig> fileConfigConfigurationUpdater) : base(templateId, outputTarget)
         {
             _sourcePath = sourcePath;
             _overwriteBehaviour = overwriteBehaviour;
+            _fileConfigConfigurationUpdater = fileConfigConfigurationUpdater;
             _replacements = replacements ?? new Dictionary<string, string>
             {
                 ["ApplicationName"] = outputTarget.ApplicationName(),
@@ -107,9 +138,9 @@ namespace Intent.Modules.Common.Templates.StaticContent
         /// <inheritdoc />
         public override ITemplateFileConfig GetTemplateFileConfig()
         {
-            var config = new TemplateFileConfig(
+            ITemplateFileConfig config = new TemplateFileConfig(
                 fileName: Path.GetFileNameWithoutExtension(_relativeOutputPath),
-                fileExtension: Path.GetExtension(_relativeOutputPath)?.TrimStart('.') ?? string.Empty,
+                fileExtension: Path.GetExtension(_relativeOutputPath).TrimStart('.') ?? string.Empty,
                 relativeLocation: Path.GetDirectoryName(_relativeOutputPath),
                 overwriteBehaviour: _overwriteBehaviour);
 
@@ -117,6 +148,8 @@ namespace Intent.Modules.Common.Templates.StaticContent
             {
                 config.CustomMetadata.Add("RelativeOutputPathPrefix", _relativeOutputPathPrefix);
             }
+
+            config = _fileConfigConfigurationUpdater(config, this);
 
             return config;
         }