+++ Modules/Intent.Modules.Common/release-notes.md
@@ -1,6 +1,7 @@
 ### Version 3.9.0
 
 - Feature: Added support for filtering template instances for registrations.
+- Improvement: A `UpdateTemplateFileConfig` virtual method was added to `StaticContentTemplateRegistration` to allow updates or replacement of the `ITemplateFileConfig` to be used by the templates.
 
 ### Version 3.8.0